+++ src/ft_math/deg_to_rad.c
@@ -6,13 +6,13 @@
 /*   By: mleblanc <mleblanc@student.42quebec.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/08 22:03:38 by mleblanc          #+#    #+#             */
-/*   Updated: 2021/11/08 22:34:10 by mleblanc         ###   ########.fr       */
+/*   Updated: 2021/11/11 12:39:01 by mleblanc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-float	deg_to_rad(float degrees)
+double	deg_to_rad(double degrees)
 {
-	return (degrees * (2.0f * (float)PI) / 360.0f);
+	return (degrees * (2.0 * PI) / 360.0);
 }
